+++ crates/pyxel-core/src/lib.rs
@@ -65,8 +65,8 @@ pub use crate::key::*;
 pub use crate::music::Music;
 pub use crate::pyxel::{
     channels, colors, cursor_image, dropped_files, font_image, frame_count, height, images, init,
-    input_keys, input_text, mouse_wheel, mouse_x, mouse_y, musics, pyxel, reset_func,
-    reset_statics, screen, sounds, tilemaps, tones, width, Pyxel,
+    input_keys, input_text, mouse_wheel, mouse_x, mouse_y, musics, pyxel, quit_callback,
+    reset_callback, reset_statics, screen, sounds, tilemaps, tones, width, Pyxel,
 };
 pub use crate::settings::*;
 pub use crate::sound::{Sound, SoundEffect, SoundNote, SoundSpeed, SoundTone, SoundVolume};